refactor(authorization-server): 인증 서버 토큰 필요한 부분만 리팩토링
